requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dateutil
openai
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
